enable auto-delete

+++ src/UploadedFilesObserver.php
@@ -9,6 +9,10 @@ class UploadedFilesObserver implements ShouldHandleEventsAfterCommit
 {
     public function updated(Model $model): void
     {
+        if (! config('file-cast.auto_delete', true)) {
+            return;
+        }
+
         if (! $model->wasChanged()) {
             return;
         }
@@ -24,6 +28,10 @@ class UploadedFilesObserver implements ShouldHandleEventsAfterCommit
 
     public function deleted(Model $model)
     {
+        if (! config('file-cast.auto_delete', true)) {
+            return;
+        }
+
         $columns = $this->getFileCastedColumns($model);
         foreach ($columns as $column) {
             $model->getAttribute($column)?->delete();
